Ale/windows support

+++ run_test.py
@@ -4,7 +4,8 @@ import os
 import csv
 import sys
 import json
-import pathlib
+from pathlib import Path
+import platform
 
 """
 parametros:
@@ -22,19 +23,37 @@ def loadController(erebus_directory, controller):
     
     print(f"Loaded {controller}")
 
+def getOpenScript(world):
+    if platform.system() == "Linux":
+        return f"""#!/bin/bash
+        webots {world} --mode=fast --minimize --no-rendering
+        """
+    elif platform.system() == "Windows":
+        return f"""webots {world} --mode=fast --minimize --no-rendering"""
+    else:
+        raise OSError("OS not supported. Please use either Windows or Linux")
+
 def openWebots(world):
-    script = f"""#!/bin/bash
-    webots {world} --mode=fast --minimize --no-rendering
-    """
+    script = getOpenScript(world)
     rc = subprocess.Popen(script, shell=True)
+
+def getKillScript():
+    if platform.system() == "Linux":
+        return """#!/bin/bash
+        pkill webots
+        """
+    elif platform.system() == "Windows":
+        return """taskkill/im webots.exe /F"""
+    
+    else:
+        raise OSError("OS not supported. Please use either Windows or Linux")
+
 
 def killWebots():
-    script = """#!/bin/bash
-    pkill webots
-    """
+    script = getKillScript()
     rc = subprocess.Popen(script, shell=True)
 
-def processLogs(world, fileName, time_taken, log_directory):
+def processLogs(world: Path, fileName: Path, time_taken, log_directory: Path):
     lastLog = sorted(os.listdir(log_directory))[-1]
 
     if "gameLog" in lastLog:
@@ -55,9 +74,9 @@ def processLogs(world, fileName, time_taken, log_directory):
         with open(fileName, "a") as file:
             writer = csv.writer(file)
 
-            writer.writerow([str(world).split("/")[-1], finalScore, finalTime, time_taken])
+            writer.writerow([world.stem, finalScore, finalTime, time_taken])
 
-def testRun(world, fileName, log_directory):
+def testRun(world: Path, fileName, log_directory: Path):
     initialLogNumber = len(os.listdir(log_directory))
     newLogNumber = len(os.listdir(log_directory))
 
@@ -81,23 +100,20 @@ def testRun(world, fileName, log_directory):
     print("Processing data...")
     processLogs(world, fileName, time_taken, log_directory)
 
-def get_output_file_name(run_name, world_set_dir):
+def get_output_file_name(run_name: str, world_set_dir: Path):
     actual_time = time.strftime("%d-%m-%Y_%H-%M-%S")
 
-    formatted_world_set_name = world_set_dir.split("/")[-1]
-    formatted_world_set_name = formatted_world_set_name.replace(".txt", "")
-
-    return run_name + "_(" + formatted_world_set_name + ")_" + actual_time
+    return run_name + "_(" + world_set_dir.stem + ")_" + actual_time
 
 def make_output_file(config):
     try:
-        os.mkdir(pathlib.Path("./runs") / config["run_name"])
+        os.mkdir(Path("./runs") / config["run_name"])
     except FileExistsError:
         print("Directory already exists")
 
-    name = get_output_file_name(config["run_name"], config["world_set"]) + ".csv"
+    name = get_output_file_name(config["run_name"], Path(config["world_set"])) + ".csv"
 
-    output_file = pathlib.Path("./runs", config["run_name"], name)
+    output_file =Path("./runs", config["run_name"], name)
 
     with open(output_file, "w") as output:
         writer = csv.writer(output)
@@ -106,7 +122,7 @@ def make_output_file(config):
     return output_file
 
 def test_runs(config):
-    erebus_directory = pathlib.Path(config["erebus_directory"])
+    erebus_directory = Path(config["erebus_directory"])
 
     log_directory = erebus_directory / "game/logs/"
